requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
urllib3==2.0.3
URLObject==2.4.3
Werkzeug==2.3.6
